[Profiling] link functions to flamegraph
