+++ x-pack/plugins/profiling/public/routing/index.tsx
@@ -13,7 +13,9 @@ import { TopNFunctionSortField, topNFunctionSortFieldRt } from '../../common/fun
 import { StackTracesDisplayOption, TopNType } from '../../common/stack_traces';
 import { ComparisonMode, NormalizationMode } from '../components/normalization_menu';
 import { RedirectTo } from '../components/redirect_to';
-import { FlameGraphsView } from '../views/flame_graphs_view';
+import { FlameGraphsView } from '../views/flamegraphs';
+import { DifferentialFlameGraphsView } from '../views/flamegraphs/differential_flamegraphs';
+import { FlameGraphView } from '../views/flamegraphs/flamegraph';
 import { FunctionsView } from '../views/functions';
 import { DifferentialTopNFunctionsView } from '../views/functions/differential_topn';
 import { TopNFunctionsView } from '../views/functions/topn';
@@ -109,9 +111,14 @@ const routes = {
                     })}
                     href="/flamegraphs/flamegraph"
                   >
-                    <Outlet />
+                    <FlameGraphView />
                   </RouteBreadcrumb>
                 ),
+                params: t.type({
+                  query: t.partial({
+                    searchText: t.string,
+                  }),
+                }),
               },
               '/flamegraphs/differential': {
                 element: (
@@ -121,7 +128,7 @@ const routes = {
                     })}
                     href="/flamegraphs/differential"
                   >
-                    <Outlet />
+                    <DifferentialFlameGraphsView />
                   </RouteBreadcrumb>
                 ),
                 params: t.type({
@@ -134,19 +141,23 @@ const routes = {
                         t.literal(ComparisonMode.Absolute),
                         t.literal(ComparisonMode.Relative),
                       ]),
-                    }),
-                    t.partial({
                       normalizationMode: t.union([
                         t.literal(NormalizationMode.Scale),
                         t.literal(NormalizationMode.Time),
                       ]),
+                    }),
+                    t.partial({
                       baseline: toNumberRt,
                       comparison: toNumberRt,
+                      searchText: t.string,
                     }),
                   ]),
                 }),
                 defaults: {
                   query: {
+                    comparisonRangeFrom: 'now-15m',
+                    comparisonRangeTo: 'now',
+                    comparisonKuery: '',
                     comparisonMode: ComparisonMode.Absolute,
                     normalizationMode: NormalizationMode.Time,
                   },
