+++ x-pack/plugins/profiling/public/views/flamegraphs/flamegraph/index.tsx
@@ -1,0 +1,81 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+import { EuiFlexGroup, EuiFlexItem, EuiHorizontalRule, EuiPanel } from '@elastic/eui';
+import React, { useState } from 'react';
+import { AsyncComponent } from '../../../components/async_component';
+import { useProfilingDependencies } from '../../../components/contexts/profiling_dependencies/use_profiling_dependencies';
+import { FlameGraph } from '../../../components/flamegraph';
+import { PrimaryProfilingSearchBar } from '../../../components/profiling_app_page_template/primary_profiling_search_bar';
+import { useProfilingParams } from '../../../hooks/use_profiling_params';
+import { useProfilingRouter } from '../../../hooks/use_profiling_router';
+import { useProfilingRoutePath } from '../../../hooks/use_profiling_route_path';
+import { useTimeRange } from '../../../hooks/use_time_range';
+import { useTimeRangeAsync } from '../../../hooks/use_time_range_async';
+
+export function FlameGraphView() {
+  const {
+    query,
+    query: { rangeFrom, rangeTo, kuery, searchText },
+  } = useProfilingParams('/flamegraphs/flamegraph');
+
+  const timeRange = useTimeRange({ rangeFrom, rangeTo });
+
+  const {
+    services: { fetchElasticFlamechart },
+  } = useProfilingDependencies();
+
+  const state = useTimeRangeAsync(
+    ({ http }) => {
+      return fetchElasticFlamechart({
+        http,
+        timeFrom: timeRange.inSeconds.start,
+        timeTo: timeRange.inSeconds.end,
+        kuery,
+      });
+    },
+    [timeRange.inSeconds.start, timeRange.inSeconds.end, kuery, fetchElasticFlamechart]
+  );
+
+  const { data } = state;
+
+  const routePath = useProfilingRoutePath();
+
+  const profilingRouter = useProfilingRouter();
+
+  const [showInformationWindow, setShowInformationWindow] = useState(false);
+  function toggleShowInformationWindow() {
+    setShowInformationWindow((prev) => !prev);
+  }
+
+  function handleSearchTextChange(newSearchText: string) {
+    // @ts-expect-error Code gets too complicated to satisfy TS constraints
+    profilingRouter.push(routePath, { query: { ...query, searchText: newSearchText } });
+  }
+
+  return (
+    <EuiFlexGroup direction="column">
+      <EuiFlexItem grow={false}>
+        <EuiPanel hasShadow={false} color="subdued">
+          <PrimaryProfilingSearchBar />
+          <EuiHorizontalRule />
+        </EuiPanel>
+      </EuiFlexItem>
+      <EuiFlexItem>
+        <AsyncComponent {...state} style={{ height: '100%' }} size="xl">
+          <FlameGraph
+            id="flamechart"
+            primaryFlamegraph={data}
+            showInformationWindow={showInformationWindow}
+            toggleShowInformationWindow={toggleShowInformationWindow}
+            searchText={searchText}
+            onChangeSearchText={handleSearchTextChange}
+          />
+        </AsyncComponent>
+      </EuiFlexItem>
+    </EuiFlexGroup>
+  );
+}
